[test] run spaces tests in parallel
this mandated increasing grid resolution
